requirements.txt
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.1.0
matplotlib==3.7.0
seaborn==0.12.0
opencv-python==4.8.0
pillow==10.0.0
jupyter==1.0.0
kaggle==1.5.16
tqdm==4.66.0
pyyaml==6.0 